fix: resolve DNSLink records that point to IPNS names
Update `@helia/verified-fetch` to resolve DNSLink records correctly
when the record resolves to an IPNS name.

Fixes #991
