+++ package.json
@@ -46,7 +46,7 @@
     "@helia/delegated-routing-v1-http-api-client": "^6.0.1",
     "@helia/routers": "^5.0.2",
     "@helia/utils": "^2.4.1",
-    "@helia/verified-fetch": "^7.0.0",
+    "@helia/verified-fetch": "^7.0.1",
     "@ipld/dag-cbor": "^9.2.5",
     "@ipld/dag-json": "^10.2.5",
     "@ipld/dag-pb": "^4.1.5",
