+++ test-conformance/fixtures/constants.ts
@@ -1,1 +1,1 @@
-export const GWC_IMAGE = process.env.GWC_IMAGE ?? 'ghcr.io/ipfs/gateway-conformance:v0.10.1'
+export const GWC_IMAGE = process.env.GWC_IMAGE ?? 'ghcr.io/ipfs/gateway-conformance:v0.11.0'
